+++ praktikumsplaner-frontend/frontend/src/api/NWKService.ts
