+++ praktikumsplaner-frontend/frontend/src/views/MeldungPraktikumsstelle/MeldungAusbildung.vue
@@ -212,7 +212,7 @@ import { Ausbildungsjahr } from "@/types/Ausbildungsjahr";
 import { Referat } from "@/types/Referat";
 import { YesNoBoolean } from "@/types/YesNoBoolean";
 import { Dringlichkeit } from "@/types/Dringlichkeit";
-import MeldungService from "@/api/MeldungService";
+import MeldungService from "@/api/PraktikumsstellenService";
 import router from "@/router";
 import { useHeaderStore } from "@/stores/header";
